+++ tardis/io/model/readers/generic_readers.py
@@ -1,24 +1,24 @@
 from astropy import units as u
+from typing import Any, Tuple
 from numpy import recfromtxt
 import pandas as pd
 from radioactivedecay import Nuclide
 from radioactivedecay.utils import Z_DICT, elem_to_Z
-
+from pathlib import Path
 
 import numpy as np
-from tardis.io.model.readers.cmfgen import (
-    read_cmfgen_composition,
-    read_cmfgen_density,
-)
+from tardis.io.model.readers.util import read_csv_isotope_abundances
 
 from tardis.util.base import parse_quantity
+
 
 class ConfigurationError(Exception):
     pass
 
 
-
-def read_simple_ascii_density(fname: str):
+def read_simple_ascii_density(
+    fname: Any,
+) -> Tuple[u.Quantity, u.Quantity, u.Quantity]:
     """
     Reading a density file of the following structure (example; lines starting with a hash will be ignored):
     The first density describes the mean density in the center of the model and is not used.
@@ -112,149 +112,6 @@ def read_artis_density(fname):
     return time_of_model, velocity, mean_density
 
 
-
-
-def read_density_file(filename, filetype):
-    """
-    read different density file formats
-
-    Parameters
-    ----------
-    filename : str
-        filename or path of the density file
-    filetype : str
-        type of the density file
-
-    Returns
-    -------
-    time_of_model : astropy.units.Quantity
-        time at which the model is valid
-    velocity : np.ndarray
-        the array containing the velocities
-    unscaled_mean_densities : np.ndarray
-        the array containing the densities
-    """
-    file_parsers = {
-        "artis": read_artis_density,
-        "simple_ascii": read_simple_ascii_density,
-        "cmfgen_model": read_cmfgen_density,
-    }
-
-    electron_densities = None
-    temperature = None
-    if filetype == "cmfgen_model":
-        (
-            time_of_model,
-            velocity,
-            unscaled_mean_densities,
-            electron_densities,
-            temperature,
-        ) = read_cmfgen_density(filename)
-    else:
-        (time_of_model, velocity, unscaled_mean_densities) = file_parsers[
-            filetype
-        ](filename)
-
-    v_inner = velocity[:-1]
-    v_outer = velocity[1:]
-
-    invalid_volume_mask = (v_outer - v_inner) <= 0
-    if invalid_volume_mask.sum() > 0:
-        message = "\n".join(
-            [
-                f"cell {i:d}: v_inner {v_inner_i:s}, v_outer " f"{v_outer_i:s}"
-                for i, v_inner_i, v_outer_i in zip(
-                    np.arange(len(v_outer))[invalid_volume_mask],
-                    v_inner[invalid_volume_mask],
-                    v_outer[invalid_volume_mask],
-                )
-            ]
-        )
-        raise ConfigurationError(
-            "Invalid volume of following cell(s):\n" f"{message:s}"
-        )
-
-    return (
-        time_of_model,
-        velocity,
-        unscaled_mean_densities,
-        electron_densities,
-        temperature,
-    )
-
-
-def read_csv_isotope_abundances(
-    fname, delimiter=r"\s+", skip_columns=0, skip_rows=[1]
-):
-    """
-    A generic parser for a TARDIS composition stored as a CSV file
-
-    The parser can read in both elemental and isotopic abundances. The first
-    column is always expected to contain a running index, labelling the grid
-    cells. The parser also allows for additional information to be stored in
-    the first skip_columns columns. These will be ignored if skip_columns > 0.
-    Note that the first column, containing the cell index is not taken into
-    account here.
-
-    Specific header lines can be skipped by the skip_rows keyword argument
-
-    It is expected that the first row of the date block (after skipping the
-    rows specified in skip_rows) specifies the different elements and isotopes.
-    Each row after contains the composition in the corresponding grid shell.
-    The first composition row describes the composition of the photosphere and
-    is essentially ignored (for the default value of skip_rows).
-
-    Example:
-
-    Index C   O   Ni56
-    0     1   1   1
-    1     0.4 0.3 0.2
-
-    Parameters
-    ----------
-    fname : str
-        filename or path with filename
-
-    Returns
-    -------
-    index : np.ndarray
-    abundances : pandas.DataFrame
-    isotope_abundance : pandas.MultiIndex
-    """
-
-    df = pd.read_csv(
-        fname, comment="#", sep=delimiter, skiprows=skip_rows, index_col=0
-    )
-    df = df.transpose()
-
-    abundance = pd.DataFrame(
-        columns=np.arange(df.shape[1]),
-        index=pd.Index([], name="atomic_number"),
-        dtype=np.float64,
-    )
-
-    isotope_index = pd.MultiIndex(
-        [[]] * 2, [[]] * 2, names=["atomic_number", "mass_number"]
-    )
-    isotope_abundance = pd.DataFrame(
-        columns=np.arange(df.shape[1]), index=isotope_index, dtype=np.float64
-    )
-
-    for element_symbol_string in df.index[skip_columns:]:
-        if element_symbol_string in Z_DICT.values():
-            z = elem_to_Z(element_symbol_string)
-            abundance.loc[z, :] = df.loc[element_symbol_string].tolist()
-        else:
-            nuc = Nuclide(element_symbol_string)
-            z = nuc.Z
-            mass_no = nuc.A
-            isotope_abundance.loc[(z, mass_no), :] = df.loc[
-                element_symbol_string
-            ].tolist()
-
-    return abundance.index, abundance, isotope_abundance
-
-
 def read_csv_composition(fname, delimiter=r"\s+"):
     """Read composition from a simple CSV file
 
@@ -303,54 +160,6 @@ def read_simple_ascii_abundances(fname):
     )
 
     return index, abundances
-
-
-def read_abundances_file(
-    abundance_filename,
-    abundance_filetype,
-    inner_boundary_index=None,
-    outer_boundary_index=None,
-):
-    """
-    read different density file formats
-
-    Parameters
-    ----------
-    abundance_filename : str
-        filename or path of the density file
-    abundance_filetype : str
-        type of the density file
-    inner_boundary_index : int
-        index of the inner shell, default None
-    outer_boundary_index : int
-        index of the outer shell, default None
-    """
-
-    file_parsers = {
-        "simple_ascii": read_simple_ascii_abundances,
-        "artis": read_simple_ascii_abundances,
-        "cmfgen_model": read_cmfgen_composition,
-        "custom_composition": read_csv_composition,
-    }
-
-    isotope_abundance = pd.DataFrame()
-    if abundance_filetype in ["cmfgen_model", "custom_composition"]:
-        index, abundances, isotope_abundance = file_parsers[abundance_filetype](
-            abundance_filename
-        )
-    else:
-        index, abundances = file_parsers[abundance_filetype](abundance_filename)
-
-    if outer_boundary_index is not None:
-        outer_boundary_index_m1 = outer_boundary_index - 1
-    else:
-        outer_boundary_index_m1 = None
-    index = index[inner_boundary_index:outer_boundary_index]
-    abundances = abundances.loc[
-        :, slice(inner_boundary_index, outer_boundary_index_m1)
-    ]
-    abundances.columns = np.arange(len(abundances.columns))
-    return index, abundances, isotope_abundance
 
 
 def read_uniform_abundances(abundances_section, no_of_shells):
