+++ tardis/io/model/readers/tests/data/arepo_cone_reference_model.csvy
