+++ tardis/io/model/readers/tests/data/artis_model.dat
