+++ tardis/io/model/readers/tests/data/invalid_artis_model.dat
