+++ tardis/io/model/readers/tests/data/mesa.stella.dat
