+++ tardis/io/model/readers/tests/data/tardis_configv1_artis_density_v_slice.yml
