+++ tardis/io/model/readers/tests/data/tardis_configv1_density_exponential_nebular.yml
