+++ tardis/io/model/readers/tests/data/tardis_configv1_nlte.yml
