+++ tardis/io/model/readers/tests/test_stella_reader.py
